+++ test/libsolidity/ASTJSON/source_location.sol
@@ -1,3 +1,3 @@
-contract C { function f() { var x = 2; x++; } }
+contract C { function f() { uint x = 2; x++; } }
 
 // ----