+++ test/libsolidity/syntaxTests/nameAndTypeResolution/invalidArgs/creating_struct_members_skipped.sol
@@ -2,8 +2,8 @@ contract C {
     struct S { uint a; uint b; mapping(uint=>uint) c; }
 
     function f() public {
-        S memory s = S({a: 1});
+        S({a: 1});
     }
 }
 // ----
-// TypeError 9515: (117-126): Struct containing a (nested) mapping cannot be constructed.
+// TypeError 9515: (104-113): Struct containing a (nested) mapping cannot be constructed.